Divided the RequestHandler into multiple files.
One file for the parsing, one file for the execution, one file for the
redirection and the main file for the rest
